Create README.md
Developed a custom logistic regression model to predict the species of Iris flowers. This project involves creating a logistic regression algorithm from scratch and applying it to the well-known Iris dataset for species classification. The goal is to gain a deeper understanding of logistic regression principles and enhance skills in machine learning implementation.